+++ CHANGELOG.md
@@ -19,7 +19,8 @@ All notable changes to this project will be documented in this file.
 - Script to create transfer bundles
 - CronEventEmitter, which emits fake Audit Log Events on a set interval
 - Observable Sensemaker, which processes all observables every 15 minutes
-  - Implemented geofence logic, other observable types to be implemented
+- Implemented geofence logic, other observable types to be implemented
+- OpenTelemetry integration for queue processing time and event counting metrics
 
 ### Changed
 - Updated Inference Sensemaker to only accept Observation Audit Log Events